Extract validateStringToken into shared tokens module
